deepfake_detection (2)/backend/requirements.txt
# Core Dependencies
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# Deep Learning Models
efficientnet-pytorch==0.7.1
timm==0.9.12
transformers==4.36.2
einops==0.7.0

# Video & Audio Processing
opencv-python==4.10.0
moviepy==1.0.3
ffmpeg-python==0.2.0
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1

# Image Processing
Pillow==10.2.0
scikit-image==0.22.0
albumentations==1.3.1

# Machine Learning Utils
scikit-learn==1.4.0
tensorboard==2.15.1
onnx==1.15.0
onnxruntime==1.16.3

# Performance Optimization
numba==0.58.1
ray==2.9.0
dask==2024.1.1
cachier==2.1.1
joblib==1.3.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# API & Networking
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
websockets==12.0
httpx==0.26.0

# Caching & Resource Management
cachetools==5.3.2
python-memcached==1.59
psutil==5.9.8
memory-profiler==0.61.0

# Validation & Type Checking
pydantic==2.6.0
typeguard==4.1.5
mypy==1.8.0
types-requests==2.31.0
types-PyYAML==6.0.12

# Configuration
python-dotenv==1.0.1
PyYAML==6.0.1
configargparse==1.7
hydra-core==1.3.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
hypothesis==6.92.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
opentelemetry-api==1.21.0
sentry-sdk==1.39.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
pdoc3==0.10.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Security
cryptography==42.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Optional: GPU Support Dependencies
cupy-cuda12x==12.3.0; platform_system == "Linux"
torch-tensorrt==1.5.0; platform_system == "Linux"

# Development Tools
jupyterlab==4.0.11
ipython==8.20.0
ipdb==0.13.13
rich==13.7.0